+++ apps/src/music/blockly/MusicBlocklyWorkspace.js
@@ -1,7 +1,7 @@
 import CustomMarshalingInterpreter from '../../lib/tools/jsinterpreter/CustomMarshalingInterpreter';
 import {BlockTypes} from './blockTypes';
 import {MUSIC_BLOCKS} from './musicBlocks';
-import {musicLabDarkTheme} from './themes';
+import musicLabDarkTheme from '@cdo/apps/blockly/themes/musicLabDark';
 import {getToolbox} from './toolbox';
 import FieldSounds from './FieldSounds';
 import {getBlockMode} from '../appConfig';
